refactor: relocate core analysis types
