+++ libapp/AnalysisRunner.h
@@ -18,7 +18,9 @@
 #include "AnalysisLogger.h"
 #include "AnalysisPluginManager.h"
 #include "AnalysisResult.h"
-#include "AnalysisTypes.h"
+#include "VariableResult.h"
+#include "SampleDataset.h"
+#include "RegionAnalysis.h"
 #include "DataProcessor.h"
 #include "HistogramBooker.h"
 #include "ISampleProcessor.h"
