+++ libdata/SampleDataset.h
@@ -1,0 +1,28 @@
+#ifndef SAMPLE_DATASET_H
+#define SAMPLE_DATASET_H
+
+#include <unordered_map>
+
+#include "ROOT/RDataFrame.hxx"
+
+#include "KeyTypes.h"
+#include "SampleTypes.h"
+
+namespace analysis {
+
+struct SampleDataset {
+    SampleOrigin origin_;
+    AnalysisRole role_;
+    mutable ROOT::RDF::RNode dataframe_;
+};
+
+struct SampleDatasetGroup {
+    SampleDataset nominal_;
+    std::unordered_map<SampleVariation, SampleDataset> variations_;
+};
+
+using SampleDatasetGroupMap = std::unordered_map<SampleKey, SampleDatasetGroup>;
+
+} // namespace analysis
+
+#endif
