+++ libplot/UnstackedHistogramPlot.h
@@ -19,7 +19,7 @@
 #include "TLine.h"
 #include "TPad.h"
 
-#include "AnalysisTypes.h"
+#include "VariableResult.h"
 #include "HistogramCut.h"
 #include "IHistogramPlot.h"
 #include "RegionAnalysis.h"
